+++ source/extensions/omni.isaac.orbit/test/sim/test_spawn_from_files.py
@@ -69,7 +69,7 @@ class TestSpawningFromFiles(unittest.TestCase):
             extension_path = get_extension_path_from_name("omni.importer.urdf")
         # Spawn franka from URDF
         cfg = sim_utils.UrdfFileCfg(
-            urdf_path=f"{extension_path}/data/urdf/robots/franka_description/robots/panda_arm_hand.urdf", fix_base=True
+            asset_path=f"{extension_path}/data/urdf/robots/franka_description/robots/panda_arm_hand.urdf", fix_base=True
         )
         prim = cfg.func("/World/Franka", cfg)
         # Check validity